server/requirements.txt
flask
flask-cors
google-generativeai
python-dotenv
openai
google-cloud-speech
google-cloud-texttospeech
googlemaps
geopy
pydantic
requests
aiohttp
python-dotenv